refactor multiannotator utils
